refactor ajuste e adiciona testes iniciai
