+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/adapter/controller/AdministradorController.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Administrador;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Administrador;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.CadastrarAdministrador;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.GerenciarAdministrador;
 import org.springframework.http.ResponseEntity;
