+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/adapter/controller/ClienteController.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Cliente;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Cliente;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.CadastrarCliente;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.GerenciarCliente;
 import org.springframework.beans.factory.annotation.Autowired;
