+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/adapter/controller/PessoaController.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Pessoa;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Pessoa;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.CadastrarPessoa;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.GerenciarPessoa;
