+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/adapter/controller/ReservaController.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Reserva;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Reserva;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.CriarReserva;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase.GerenciarReserva;
 import org.springframework.beans.factory.annotation.Autowired;
