+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/adapter/repository/MongoAvaliacaoRepository.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.repository;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Avaliacao;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Avaliacao;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.AvaliacaoRepository;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
