+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/adapter/repository/MongoRestauranteRepository.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.repository;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Restaurante;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Restaurante;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.RestauranteRepository;
 
 import org.springframework.data.mongodb.repository.MongoRepository;
