+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/domain/entity/Administrador.java
@@ -1,4 +1,4 @@
-package br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain;
+package br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
@@ -9,22 +9,18 @@ import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Data
 @Document
 @JsonIdentityInfo(generator = ObjectIdGenerators.PropertyGenerator.class, property = "id")
-public class Cliente {
+public class Administrador {
 
     @Id
     private String id;
 
     @DBRef
-    private List<Reserva> reservas = new ArrayList<>(); // Inicializa a lista vazia
-
-    @DBRef
-    private List<Avaliacao> avaliacoes = new ArrayList<>(); // Inicializa a lista vazia
+    private List<Restaurante> restaurantes;
 
     @DBRef
     @Indexed(unique = true)  // Definindo o campo `pessoa` como único na base de dados
