+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/domain/entity/Avaliacao.java
@@ -1,4 +1,4 @@
-package br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain;
+package br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity;
 
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonIdentityReference;
@@ -8,11 +8,9 @@ import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.DBRef;
 import org.springframework.data.mongodb.core.mapping.Document;
 
-import java.time.LocalDateTime;
-
 @Data
-@Document(collection = "reservas")
-public class Reserva {
+@Document(collection = "avaliacoes")
+public class Avaliacao {
 
     @Id
     private String id;
@@ -27,6 +25,6 @@ public class Reserva {
     @JsonIdentityReference(alwaysAsId = true)
     private Cliente cliente;
 
-    private LocalDateTime dataHoraReserva;
-    private int numeroDePessoas;
+    private int nota;
+    private String comentario;
 }
