+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/domain/repository/RestauranteRepository.java
@@ -1,6 +1,6 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository;
 
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Restaurante;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Restaurante;
 
 import java.util.Optional;
 import java.util.List;
