+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/usecase/CadastrarRestaurante.java
@@ -1,10 +1,12 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase;
 
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller.exception.ControllerMessagingException;
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Restaurante;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Restaurante;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.RestauranteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class CadastrarRestaurante {
