+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/usecase/CriarReserva.java
@@ -2,7 +2,7 @@ package br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase;
 
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller.exception.ControllerMessagingException;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller.exception.ControllerNotFoundException;
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Reserva;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Reserva;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.ClienteRepository;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.ReservaRepository;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.RestauranteRepository;
