+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/usecase/GerenciarCliente.java
@@ -1,7 +1,8 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase;
 
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller.exception.ControllerNotFoundException;
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Cliente;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Cliente;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Pessoa;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.ClienteRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
