+++ src/main/java/br/com/fiap/pos_tech_adj/tech_challenge_fase3/usecase/GerenciarReserva.java
@@ -1,7 +1,7 @@
 package br.com.fiap.pos_tech_adj.tech_challenge_fase3.usecase;
 
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.adapter.controller.exception.ControllerNotFoundException;
-import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.Reserva;
+import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.entity.Reserva;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.ClienteRepository;
 import br.com.fiap.pos_tech_adj.tech_challenge_fase3.domain.repository.ReservaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
